Move unit tests to files they apply to
